requirements.txt
django-templated-email==0.5
django-celery-beat==1.1.1
celery==4.1.0
Django==1.11.13
mysqlclient==1.3.12
django-debug-toolbar==1.6
django-webpack-loader==0.6.0
djangorestframework==3.7.7
python-social-auth==0.3.6
configparser==3.5.0
social-auth-app-django==2.1.0
rest-framework-generic-relations==1.1.0
django-widget-tweaks==1.4.1